+++ client/package.json
@@ -9,6 +9,7 @@
     "passport": "^0.4.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.1"
   },
   "scripts": {
