+++ client/src/App.js
@@ -1,21 +1,26 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
+import React from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import Admin from "./pages/admin";
+import Business from "./pages/business";
+import CreateRequest from "./pages/createRequest";
+import Fiscal from "./pages/fiscal";
+import Initiator from "./pages/initiator";
+import Login from "./pages/login";
+import "./App.css";
 
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
-    );
-  }
-}
+const App = () => (
+  <Router>
+    <div>
+      <Switch>
+        <Route exact path="/" component={Login} />
+        <Route exact path="/request" component={CreateRequest} />
+        <Route exact path="/admin" component={Admin} />
+        <Route exact path="/business" component={Business} />
+        <Route exact path="/fiscal" component={Fiscal} />
+        <Route exact path="/initiator" component={Initiator} />
+      </Switch>
+    </div>
+  </Router>
+);
 
 export default App;
