+++ client/src/components/HelloMsg.js
@@ -4,7 +4,7 @@ class HelloMsg extends Component {
     render() {
         return (
             <div>
-                <h2>(this should be big and bold) Hello "first name" "last Name"!</h2>
+                <h2><em>Hello name!</em></h2>
             </div>
         );
     }
