+++ package.json
@@ -21,7 +21,8 @@
     "nodemon": "^1.11.0"
   },
   "dependencies": {
-    "express": "^4.15.4"
+    "express": "^4.15.4",
+    "react-router-dom": "^4.2.2"
   },
   "bugs": {
     "url": "https://github.com/Sylvrleef/Project3-Team1/issues"
